+++ compiler/component/execution/index.js
@@ -10,7 +10,6 @@ const Primative = {
 };
 
 const ExecutionFlow = require('./flow.js');
-const Structure = require('../struct.js');
 const Variable = require('../memory/variable.js');
 
 class Execution extends ExecutionFlow {
@@ -35,7 +34,6 @@ class Execution extends ExecutionFlow {
 		frag.merge(access.preamble);
 		access = access.variable;
 
-
 		// Resolve the expression
 		let expr = this.compile_expr(ast.tokens[1], access.type, true);
 		if (expr === null) {
@@ -49,6 +47,16 @@ class Execution extends ExecutionFlow {
 				`Error: Assignment type mis-match` +
 				` cannot assign ${targetType.toString()}` +
 				` to ${expr.type.toString()}`,
+				ast.ref.start, ast.ref.end
+			);
+			return null;
+		}
+
+		if (!access.isUndefined() && access.type.type.getDestructor()) {
+			this.getFile().throw(
+				`Error: Unsafe value drop\n` +
+				`  The previous value of ${access.name} was not destructed or consumed - but has now been lost\n` +
+				`  Suggest adding ${access.type.type.name}.Delete(${access.name}) before the current line`,
 				ast.ref.start, ast.ref.end
 			);
 			return null;
@@ -166,20 +174,18 @@ class Execution extends ExecutionFlow {
 			return null;
 		}
 
-		if (target.type.type.meta == "CLASS") {
-			let destructor = target.type.type.getDestructor();
-			if (destructor) {
-				if (this.ctx == destructor) {
-					this.getFile().throw(
-						`Error: Dangerous destructor, does not properly destruct all child values`,
-						ast.ref.start, ast.ref.end
-					);
-				} else {
-					this.getFile().warn(
-						`Warn: This class type has a destructor, recommend calling ${target.type.type.name}.Delete(${target.name})`,
-						ast.ref.start, ast.ref.end
-					);
-				}
+		let destructor = target.type.type.getDestructor();
+		if (destructor) {
+			if (this.ctx == destructor) {
+				this.getFile().throw(
+					`Error: Dangerous destructor, does not properly destruct all child values`,
+					ast.ref.start, ast.ref.end
+				);
+			} else {
+				this.getFile().warn(
+					`Warn: This class type has a destructor, recommend calling ${target.type.type.name}.Delete(${target.name})`,
+					ast.ref.start, ast.ref.end
+				);
 			}
 		}
 
@@ -327,33 +333,25 @@ class Execution extends ExecutionFlow {
 		if (out === null) {
 			return null;
 		}
-
 		frag.merge(out.preamble);
 
-		// Put the value into a temporary variable to destruct the non-used value=
-		if (out.type.type.represent == "void") {
+		// Ensure value destruction is preserved
+		let type = out.type.type;
+		if (type.represent == "void") {
 			frag.append(out.instruction);
 		} else {
-			let target;
-			if (out.type.type.typeSystem == "linear") {
-				target = out.instruction;
-			} else {
-				let id = new LLVM.ID();
-				frag.append(new LLVM.Set(
-					new LLVM.Name(id, false, ast.ref),
-					out.instruction,
-					ast.ref
-				));
-				target = new LLVM.Argument(
-					out.type.toLLVM(ast.ref),
-					new LLVM.Name(id.reference(), false, ast.ref),
-					ast.ref
+			let destructor = type.getDestructor();
+			if (destructor) {
+				this.getFile().throw(
+					`Error: Unhandled return value.\n` +
+					`  The return type has a destructor which is not executed for the unhandled value\n` +
+					`  Suggest putting the statement in a ${type.name}.Delete() function`,
+					ast.ref.start,
+					ast.ref.end
 				);
-			}
 
-			let temp = new Variable(out.type, "return", ast.ref);
-			temp.markUpdated(target);
-			frag.merge(temp.cleanup(ast.ref));
+				return null;
+			}
 		}
 
 		// merge any epilog of the call
@@ -430,6 +428,7 @@ class Execution extends ExecutionFlow {
 			}
 			returnType = res.type;
 			frag.merge(res.preamble);
+
 			if (returnType.type.typeSystem == "linear") {
 
 				let size = returnType.type.sizeof(ast.ref);
@@ -486,6 +485,7 @@ class Execution extends ExecutionFlow {
 			}
 		}
 
+		// Check the return type is correct
 		if (!this.returnType.match(returnType)) {
 			this.getFile().throw(
 				`Return type miss-match, expected ${this.returnType.toString()} but got ${returnType.toString()}`,
@@ -493,13 +493,14 @@ class Execution extends ExecutionFlow {
 			);
 		}
 
+
 		// Clean up the scope
-		let res = this.scope.cleanup(ast.ref);
-		if (res.error) {
-			this.getFile().throw(res.msg, res.ref.start, res.ref.end);
+		let clean = this.scope.cleanup(ast.ref);
+		if (clean.error) {
+			this.getFile().throw(clean.msg, clean.ref.start, clean.ref.end);
 			return null;
 		}
-		frag.append(res);
+		frag.append(clean);
 
 		frag.append(new LLVM.Return(inner, ast.ref.start));
 		this.returned = true;
@@ -570,6 +571,16 @@ class Execution extends ExecutionFlow {
 
 		if (!failed && this.returned == false && !this.isChild) {
 			if (this.returnType.type == Primative.types.void) {
+				// Auto generate return and cleanup for void functions
+
+				// Clean up the scope
+				let res = this.scope.cleanup(ast.ref);
+				if (res.error) {
+					this.getFile().throw(res.msg, res.ref.start, res.ref.end);
+				} else {
+					fragment.append(res);
+				}
+
 				fragment.append(new LLVM.Return(
 					new LLVM.Type("void", 0),
 					ast.ref