+++ compiler/primative/types.js
@@ -1,4 +1,5 @@
 const TypeDef = require('../component/typedef.js');
+const Type = require('../middle/type.js');
 
 let types = {
 	void: new TypeDef(null, {
@@ -66,7 +67,7 @@ let types = {
 		}
 	}, true),
 
-	string: new TypeDef(null, {
+	cstring: new TypeDef(null, {
 		tokens: [
 			{
 				type   : "name",
@@ -74,7 +75,24 @@ let types = {
 			},
 			{
 				type   : "integer",
-				tokens : "1"
+				tokens : "8"
+			}
+		],
+		ref: {
+			start: null,
+			end: null
+		}
+	}, true),
+
+	unsafe_blob: new TypeDef(null, {
+		tokens: [
+			{
+				type   : "name",
+				tokens : "i8*"
+			},
+			{
+				type   : "integer",
+				tokens : "8"
 			}
 		],
 		ref: {
@@ -135,7 +153,7 @@ types.double.cat = "float";
 types.double.signed = true;
 
 types.int   = types.i64;
-types.uint   = types.u64;
+types.uint  = types.u64;
 
 
 // Update primative types correct type system
@@ -143,8 +161,12 @@ for (let key in types) {
 	types[key].primative = true;
 	types[key].typeSystem = 'normal';
 }
-types.string.typeSystem = "linear";
-types.string.primative = true;
+types.cstring.name = "cstring";
+types.cstring.typeSystem = "linear";
+types.cstring.primative = true;
+types.unsafe_blob.name = "unsafe_blob";
+types.unsafe_blob.typeSystem = "normal";
+types.unsafe_blob.primative = true;
 
 
 module.exports = types;