+++ changelog.md
@@ -7,6 +7,10 @@
 - [x] Primitive class implementation
 - [x] Delete instruction to consume values
 - [x] Class destructors
+- [x] Class clone operations
+- [x] String class
+- [x] No longer need a return statement on a void function
+- [x] Bitwise cast (for core development only)
 
 ## Changes
 - [x] The ``uvc`` command now executes the result by default.